requirements.txt
numpy
opencv-python
albumentations
wandb
hydra-core
hydra-optuna-sweeper
matplotlib
seaborn
tqdm
segmentation-models-pytorch
torch
torchvision
torchaudio
